// types/EligibilityCheckSuccessEligible.ts
/**
 * Do not edit this file it is auto-generated by italia-utils / gen-api-models.
 * See https://github.com/teamdigitale/italia-utils
 */
/* tslint:disable */

import { FamilyMembers } from "./FamilyMembers";
import { MaxBonusAmount } from "./MaxBonusAmount";
import { MaxBonusTaxBenefit } from "./MaxBonusTaxBenefit";
import { Timestamp } from "./Timestamp";
import * as t from "io-ts";
import { NonEmptyString } from "italia-ts-commons/lib/strings";
import { enumType } from "italia-ts-commons/lib/types";

export enum StatusEnum {
  "ELIGIBLE" = "ELIGIBLE"
}

/**
 * Eligibility check succeeded, we can proceed and send the eligibility status.
 * Amounts are left undefined when status is INELIGIBLE
 *
 */

// required attributes
const EligibilityCheckSuccessEligibleR = t.interface({
  id: NonEmptyString,

  status: enumType<StatusEnum>(StatusEnum, "status"),

  familyMembers: FamilyMembers,

  maxAmount: MaxBonusAmount,

  maxTaxBenefit: MaxBonusTaxBenefit,

  validBefore: Timestamp
});

// optional attributes
const EligibilityCheckSuccessEligibleO = t.partial({});

export const EligibilityCheckSuccessEligible = t.exact(
  t.intersection(
    [EligibilityCheckSuccessEligibleR, EligibilityCheckSuccessEligibleO],
    "EligibilityCheckSuccessEligible"
  )
);

export type EligibilityCheckSuccessEligible = t.TypeOf<
  typeof EligibilityCheckSuccessEligible
>;