// types/EligibilityCheckFailure.ts
/**
 * Do not edit this file it is auto-generated by italia-utils / gen-api-models.
 * See https://github.com/teamdigitale/italia-utils
 */
/* tslint:disable */

import * as t from "io-ts";
import { NonEmptyString } from "italia-ts-commons/lib/strings";
import { enumType } from "italia-ts-commons/lib/types";

export enum ErrorEnum {
  "INVALID_REQUEST" = "INVALID_REQUEST",

  "INTERNAL_ERROR" = "INTERNAL_ERROR",

  "DATA_NOT_FOUND" = "DATA_NOT_FOUND",

  "DATABASE_OFFLINE" = "DATABASE_OFFLINE"
}

/**
 * Eligibility check failed because of some INPS internal error
 */

// required attributes
const EligibilityCheckFailureR = t.interface({
  id: NonEmptyString,

  error: enumType<ErrorEnum>(ErrorEnum, "error"),

  errorDescription: t.string
});

// optional attributes
const EligibilityCheckFailureO = t.partial({});

export const EligibilityCheckFailure = t.exact(
  t.intersection(
    [EligibilityCheckFailureR, EligibilityCheckFailureO],
    "EligibilityCheckFailure"
  )
);

export type EligibilityCheckFailure = t.TypeOf<typeof EligibilityCheckFailure>;